add bet list and detail screens
