+++ lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:shot_bets/HomeScreen.dart';
 
 void main() => runApp(MyApp());
 
@@ -7,61 +8,11 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'An app to track shot bets among groups.',
+      title: 'An app to track shot bets between friends.',
       theme: ThemeData(
         primarySwatch: Colors.amber,
       ),
-      home: HomeScreen()
+      home: HomeScreen(),
     );
   }
-}
-
-class _HomeScreenState extends State<HomeScreen> {
-  int _selectedIndex = 0;
-  final _widgetOptions = [
-    Text('Index 0: Home'),
-    Text('Index 1: Bets'),
-    Text('Index 2: Settings'),
-  ];
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('Shot Bets!'),
-        actions: <Widget>[
-          IconButton(
-            icon: Icon(Icons.add),
-            tooltip: 'Add a new bet',
-            onPressed: () {
-              print('open new bet screen');
-            },
-          ),
-        ],
-      ),
-      body: Center(
-        child: _widgetOptions.elementAt(_selectedIndex),
-      ),
-      bottomNavigationBar:
-        BottomNavigationBar(
-          items: <BottomNavigationBarItem>[
-            BottomNavigationBarItem(icon: Icon(Icons.home), title: Text("Home")),
-            BottomNavigationBarItem(icon: Icon(Icons.local_drink), title: Text("Bets")),
-            BottomNavigationBarItem(icon: Icon(Icons.settings), title: Text("Settings")),
-          ],
-          currentIndex: _selectedIndex,
-          onTap: (index) =>
-            setState(() {
-              _selectedIndex = index;
-            }),
-        ),
-    );
-  }
-}
-
-class HomeScreen extends StatefulWidget {
-  HomeScreen({Key key}) : super(key: key);
-
-  @override
-  _HomeScreenState createState() => _HomeScreenState();
 }
